- Late fabs/fabsf() fixing.

+++ src/synthv1_lv2.cpp
@@ -184,7 +184,7 @@ void synthv1_lv2::run ( uint32_t nframes )
 						}
 						if (synthv1::paramValue(synthv1::DEL1_BPMSYNC) > 0.0f) {
 							const float bpm = synthv1::paramValue(synthv1::DEL1_BPM);
-							if (bpm > 0.0f && ::fabs(bpm_host - bpm) > 0.01f)
+							if (bpm > 0.0f && ::fabsf(bpm_host - bpm) > 0.01f)
 								synthv1::setParamValue(synthv1::DEL1_BPM, bpm_host);
 						}
 					}
